+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Nexus storage change ([#1341](https://github.com/opendevstack/ods-core/issues/1341))
 - Update PVC migration script, adding threads to rsync execution ([#1345](https://github.com/opendevstack/ods-core/pull/1345))
 - Update Aqua cli to 760 ([#1344](https://github.com/opendevstack/ods-core/pull/1344))
+- Adapted Sonarqube server configuration to make projects private and have custom gate ([#1347](https://github.com/opendevstack/ods-core/pull/1347))
 
 
 ### Fixed